+++ iceberg/streams/iceberg/file_reader_builder.h
@@ -7,23 +7,14 @@
 #include <cstdint>
 
 #include <iceberg/common/rg_metadata.h>
-
-#include <map>
-#include <memory>
-
 #include <parquet/metadata.h>
-
-#include <string>
-#include <utility>
-#include <vector>
 
 #include "iceberg/common/batch.h"
 #include "iceberg/common/fs/file_reader_provider.h"
 #include "iceberg/common/logger.h"
 #include "iceberg/common/rg_metadata.h"
 #include "iceberg/common/selection_vector.h"
-#include "iceberg/streams/arrow/file_reader.h"
-#include "iceberg/streams/arrow/projection_stream.h"
+#include "iceberg/streams/arrow/batch_with_row_number.h"
 #include "iceberg/streams/iceberg/data_entries_meta_stream.h"
 #include "iceberg/streams/iceberg/data_scan.h"
 #include "iceberg/streams/iceberg/equality_delete_applier.h"
@@ -31,7 +22,6 @@
 #include "iceberg/streams/iceberg/mapper.h"
 #include "iceberg/streams/iceberg/plan.h"
 #include "iceberg/streams/iceberg/row_group_filter.h"
-#include "parquet/arrow/reader.h"
 #include "parquet/type_fwd.h"
 
 namespace iceberg {
@@ -78,105 +68,7 @@ class FileReaderBuilder : public DataScanner::IIcebergStreamBuilder {
     Ensure(file_reader_provider_ != nullptr, std::string(__PRETTY_FUNCTION__) + ": file_reader_provider is nullptr");
   }
 
-  IcebergStreamPtr Build(const AnnotatedDataPath& annotated_data_path) override {
-    auto arrow_reader = MakeArrowReader(annotated_data_path.GetPath());
-    Ensure(arrow_reader != nullptr, std::string(__PRETTY_FUNCTION__) + ": arrow_reader is nullptr");
-
-    if (logger_) {
-      logger_->Log(std::to_string(1), "metrics:data:files_read");
-    }
-
-    auto parquet_reader = arrow_reader->parquet_reader();
-    Ensure(parquet_reader != nullptr, std::string(__PRETTY_FUNCTION__) + ": parquet_reader is nullptr");
-
-    std::shared_ptr<const parquet::FileMetaData> metadata = parquet_reader->metadata();
-    Ensure(metadata != nullptr, std::string(__PRETTY_FUNCTION__) + ": metadata is nullptr");
-
-    const std::vector<int> field_ids_required = [&]() {
-      auto field_ids_for_equality_delete = equality_deletes_->GetFieldIds(annotated_data_path.GetPartitionLayer());
-
-      if (logger_) {
-        logger_->Log(std::to_string(field_ids_for_equality_delete.size()), "metrics:data:columns_equality_delete");
-      }
-
-      std::vector<int> result = field_ids_to_retrieve_;
-      for (const int f : field_ids_for_equality_delete) {
-        result.emplace_back(f);
-      }
-
-      std::sort(result.begin(), result.end());
-      result.erase(std::unique(result.begin(), result.end()), result.end());
-
-      if (logger_) {
-        logger_->Log(std::to_string(static_cast<int>(result.size()) - static_cast<int>(field_ids_to_retrieve_.size())),
-                     "metrics:data:columns_only_equality_delete");
-      }
-
-      return result;
-    }();
-
-    const auto cols = GetColumnPositionsToRetrieveByFieldIds(*metadata, field_ids_required);
-
-    std::vector<int> col_positions = [&]() {
-      std::vector<int> result;
-      for (const auto& col : cols) {
-        if (col.column_position.has_value()) {
-          result.emplace_back(col.column_position.value());
-        }
-      }
-      return result;
-    }();
-
-    const std::vector<int> row_groups_matching_offsets =
-        GetRowGroupsToRetrieve(*metadata, annotated_data_path.GetSegments());
-    int64_t planned_row_groups = row_groups_matching_offsets.size();
-    int64_t skipped_row_groups = 0;
-
-    const std::vector<int> matching_row_groups = [&]() {
-      if (!rg_filter_) {
-        return row_groups_matching_offsets;
-      }
-
-      std::vector<int> result;
-      for (int rg_ind : row_groups_matching_offsets) {
-        std::shared_ptr<const parquet::RowGroupMetaData> rg_meta = metadata->RowGroup(rg_ind);
-        Ensure(rg_meta != nullptr, std::string(__PRETTY_FUNCTION__) + ": rg_meta is nullptr");
-
-        if (!rg_filter_->CanSkipRowGroup(*rg_meta)) {
-          result.emplace_back(rg_ind);
-        } else {
-          ++skipped_row_groups;
-        }
-      }
-
-      return result;
-    }();
-
-    if (logger_) {
-      logger_->Log(std::to_string(planned_row_groups), "metrics:row_groups:planned");
-      logger_->Log(std::to_string(skipped_row_groups), "metrics:row_groups:skipped");
-
-      logger_->Log(std::to_string(col_positions.size()), "metrics:data:columns_read");
-    }
-
-    StreamPtr<ArrowBatchWithRowPosition> result = std::make_shared<ParquetFileReader>(
-        std::shared_ptr<parquet::arrow::FileReader>(arrow_reader), matching_row_groups, col_positions, logger_);
-
-    const std::map<std::string, std::string> parquet_name_to_result_name = [&]() {
-      std::map<std::string, std::string> result;
-      for (const auto& col : cols) {
-        if (col.column_position.has_value()) {
-          result[col.parquet_name] = col.result_name;
-        }
-      }
-
-      return result;
-    }();
-
-    result = std::make_shared<ProjectionStream>(parquet_name_to_result_name, result);
-
-    return std::make_shared<ArrowStreamWrapper>(result, annotated_data_path.GetPartitionLayerFile());
-  }
+  IcebergStreamPtr Build(const AnnotatedDataPath& annotated_data_path) override;
 
  private:
   std::shared_ptr<parquet::arrow::FileReader> MakeArrowReader(const std::string& path) {
@@ -262,6 +154,11 @@ class FileReaderBuilder : public DataScanner::IIcebergStreamBuilder {
     }
     return row_groups_to_process;
   }
+
+  StreamPtr<ArrowBatchWithRowPosition> MakeFinalStream(std::shared_ptr<parquet::arrow::FileReader> arrow_reader,
+                                                       const std::vector<int>& matching_row_groups,
+                                                       std::shared_ptr<const parquet::FileMetaData> metadata,
+                                                       const std::vector<int>& field_ids);
 
   const std::vector<ColumnInfo> columns_to_retrieve_;
   const std::vector<int> field_ids_to_retrieve_;
